Seuil alerte done

+++ src/Application/DTOs/SeuilAlerte/SeuilAlerteDto.cs
@@ -9,7 +9,7 @@ public class SeuilAlerteDto
     public TypeSeuil TypeSeuil { get; set; }
     public TypeAlerte TypeAlerte { get; set; }
 
-    public double Valeur { get; set; }
+    public decimal Valeur { get; set; }
     public bool EstActif { get; set; }
 
     public DateTime DateCreation { get; set; }
